Waksman routing implementation ; updated install
A Waksman routing implementation replaces the previous Benes routing
implementation for RAM operations. This provides better efficiency
for programs with non-power-of-two number of RAM operations.

We've revamped the install procedure to make it somewhat easier.
This includes a new install script (in the thirdparty repo),
and updated instructions for Fedora and Debian (and, presumably,
related distributions).